fix(tui): address usage review feedback
