+++ codex-rs/tui/src/chatwidget/settings.rs
@@ -227,14 +227,16 @@ impl ChatWidget {
         status_account_display: Option<StatusAccountDisplay>,
         plan_type: Option<PlanType>,
         has_chatgpt_account: bool,
+        has_codex_backend_auth: bool,
     ) {
         self.status_account_display = status_account_display;
         self.plan_type = plan_type;
         self.has_chatgpt_account = has_chatgpt_account;
+        self.has_codex_backend_auth = has_codex_backend_auth;
         self.bottom_pane
             .set_connectors_enabled(self.connectors_enabled());
         self.bottom_pane
-            .set_token_activity_command_enabled(has_chatgpt_account);
+            .set_token_activity_command_enabled(has_codex_backend_auth);
     }
 
     pub(crate) fn set_realtime_audio_device(
