+++ codex-rs/tui/src/chatwidget/slash_dispatch.rs
@@ -960,7 +960,7 @@ impl ChatWidget {
             collaboration_modes_enabled: self.collaboration_modes_enabled(),
             connectors_enabled: self.connectors_enabled(),
             plugins_command_enabled: self.config.features.enabled(Feature::Plugins),
-            token_activity_command_enabled: self.has_chatgpt_account,
+            token_activity_command_enabled: self.has_codex_backend_auth,
             goal_command_enabled: self.config.features.enabled(Feature::Goals),
             service_tier_commands_enabled: self.fast_mode_enabled(),
             personality_command_enabled: self.config.features.enabled(Feature::Personality),
@@ -972,7 +972,7 @@ impl ChatWidget {
     }
 
     fn ensure_token_activity_command_available(&mut self) -> bool {
-        if self.has_chatgpt_account {
+        if self.has_codex_backend_auth {
             return true;
         }
         self.add_error_message(USAGE_CHATGPT_LOGIN_REQUIRED.to_string());
